core/requirements.txt
names
matplotlib
networkx==1.9.1
PyYAML==3.11
click==4.0
